Fixes underscore treated as wildcard in emails
In a project we are using this module, and having issues with users unable to sign up if their email address has an underscore in it and closely matches an existing email in the system. After a bit of digging, I found that the LIKE operator in Postgres and MySQL uses underscore as a single character wildcard.

It seems the SilverStripe ORM's 'nocase' modifier must use a LIKE query under the hood, at least in MySQL backends.

Email addresses like 'me2@me.com' and 'm_2@me.com' should not match, but they do with filter('Email:nocase’)

Raised issue on Silverstripe Framework: silverstripe/silverstripe-framework#10074